Remove DistanceESMDA in favor of LocalizedESMDA
Delete experimental.py and its tests.
LocalizedESMDA provides equivalent functionality through a callback-based
API. Remove calc_max_number_of_layers_per_batch_for_distance_localization
and the psutil dependency from utils.py.

Add a distance-based localization test to test_esmda_localized.py
verifying that spatially-decaying callbacks leave distant parameters
unchanged.
