+++ pyproject.toml
@@ -15,7 +15,7 @@ classifiers=[
     "Programming Language :: Python :: 3.13",
     "Programming Language :: Python :: 3.14",
 ]
-dependencies = ["numpy", "scipy", "joblib", "psutil"]
+dependencies = ["numpy", "scipy", "joblib"]
 dynamic=["version"] # setuptools_scm extracts Python package versions from git
 
 authors = [
